+++ bebytes/tests/direct_buffer_writing.rs
@@ -1,5 +1,3 @@
-#![cfg(feature = "bytes")]
-
 use bebytes::BeBytes;
 use bytes::BytesMut;